+++ src/assets/config/config.ts
@@ -331,7 +331,11 @@ export const config: Config = {
         indexPersons: true,
         indexPlaces: true,
         indexWorks: false,
-        search: true
+        search: true,
+        cookiePolicy: false,
+        privacyPolicy: false,
+        termsOfUse: false,
+        accessibilityStatement: false,
       },
       defaultExpanded: false,
       ungroupArticles: false